+++ MySql/scripts/admin/change_suggest_id.js
@@ -1,14 +1,8 @@
 const db = require("../../config/create_connection");
 
-async function VerifyUser({
-  response,
-  user_table_id,
-  matric_number,
-  firstname,
-  hardware_id,
-}) {
+async function ChangeSuggestId({ response, next_suggest_id }) {
   try {
-    db.connect(async (err) => {
+    db.connect((err) => {
       if (err) {
         response.status(400).send({
           success: false,
@@ -17,8 +11,7 @@ async function VerifyUser({
         });
       } else {
         db.query(
-          `INSERT INTO verified_users (user_table_id, matric_number) VALUES (?, ?);`,
-          [Number(user_table_id), matric_number.toLowerCase()],
+          `UPDATE suggest_id SET suggest_id = ${next_suggest_id} WHERE id = 1;`,
           (err, result) => {
             if (err) {
               response.status(400).send({
@@ -27,16 +20,11 @@ async function VerifyUser({
                 data: err,
               });
             } else {
-              response.status(200).send({
+              response.status(201).send({
                 success: true,
-                message: `Student name: ${firstname} with Matric Number: ${matric_number} was successfully verified`,
+                message: `Next suggested ID updated to - ${next_suggest_id}`,
                 data: {
-                  verified_id: hardware_id,
-                  student_data: {
-                    user_id: hardware_id,
-                    student_name: firstname,
-                    matric_no: matric_number,
-                  },
+                  next_suggest_id,
                 },
               });
             }
@@ -52,4 +40,4 @@ async function VerifyUser({
     });
   }
 }
-exports.VerifyUser = VerifyUser;
+exports.ChangeSuggestId = ChangeSuggestId;